pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "docint"
version = "0.1.0"
description = "Production document intelligence pipeline — structured extraction from PDFs, DOCX, XLSX, and images"
requires-python = ">=3.11"
dependencies = [
    # LLM
    "anthropic>=0.40.0",
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    # Parsing
    "pdfplumber>=0.11.0",
    "PyMuPDF>=1.24.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    # Storage
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Schemas & validation
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "PyYAML>=6.0.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
